Make release script: better GPG handling
Fixed gpg-agent usage so that it detects the useful environment variable
rather than the process name (which is useless without the environment
variable). Does a dummy gpg operation at the start to cause a password
prompt, which should cause gpg-agent to cache the password.

Tweak the gpg-agent configuration example (also used in the Vagrantfile)
to lower the ridiculously long timeouts.