remove the properties section from this file. they are all pulled in from the parent (brooklyn-server/pom.xml).
(except for jackson which we don't want to declare as we don't use it.)

sidebar comment, i now tend to the opinion that we *don't* want this.  downstream projects should define their own workflow.
the main value is ensuring versions are compatible but that is something we can/should handle a different way.
(simple way, just by running the license-audit plugin and eyeball anything with multiple versions...)