archetype’s catalog.bom set up as 0.1.0-SNAPSHOT
Previously was 0.10.0-SNAPSHOT, but this is the version of the user’s
project rather than brooklyn.

Also removes the apache header because this is a starting point for
a user to create their own projects (which will not necessarily be
apache licensed). This starting file is extremely basic; it is a file
without any degree of creativity.