switch vagrant boxes to bento images
Prior to this commit running the vagrant getting started environment
would fall back to using rsync for making the /vagrant share available
in the guest OS due to the `centos/7` box not including virtualbox guest
additions (resulting in the absense of the vboxfs drivers). The
`centos/X` boxes do not following current best practices for virtualbox
providers.

This commit switches to the community managed bentos centos-7.3 box
which follows best practices and includes the vbox guest additions. See
- https://github.com/chef/bento