Update merge script to support optional non-master branch.
For convenience merging commits into branches apart from master.

Usage:

  am 588 0.10.x

will merge 588 from the 0.10.x branch of ${PR_REMOTE}.

The script does not check out the branch!  It requires you to check it out
and verifies that it is on the right branch.  Leaving it like this to enforce
a certain manualness to the procedure, for safety's sake.